requirements.txt
paho-mqtt==1.5.0
RPi.GPIO==0.7.0
smbus==1.1.post2
